crypto/asymmetric-cipher/src/requirements.txt
# Crypto - Asymmetric Cipher Dependencies
owiener>=0.1.0
pycryptodome>=3.18.0
gmpy2>=2.1.5
sympy>=1.12
